# web/openapi.yaml
openapi: 3.1.0
info:
  title: SkillHub API
  description: |
    SkillHub registry HTTP API — discovery, publishing, authentication, namespaces,
    stars/comments/ratings, notifications, and admin operations.

    All JSON endpoints live under `/api/v1`. Bearer tokens (`Authorization: Bearer <token>`)
    are obtained via the device flow (`/api/v1/auth/device/*`) or via a logged-in web
    session cookie (`session_token`).

    See `/skills.md` for a runnable curl-based walkthrough designed for AI agents.
  version: "1.0.0"
  license:
    name: MIT
servers:
  - url: /
    description: Same-origin
...
paths:
...
  /api/v1/plugins:
    get:
      tags: [Plugins]
      summary: List plugins (paginated)
      parameters:
        - { name: cursor,   in: query, schema: { type: string },  description: "Opaque cursor from previous nextCursor" }
        - { name: sort,     in: query, schema: { type: string, enum: [created, downloads, stars, name], default: created } }
        - { name: limit,    in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: category, in: query, schema: { type: string } }
      responses:
        "200":
          description: Paginated plugin list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:       { type: array, items: { $ref: "#/components/schemas/Plugin" } }
                  nextCursor: { type: string, nullable: true }
    post:
      tags: [Plugins]
      summary: Publish a new plugin version (multipart)
      description: |
        Upload a plugin bundle as multipart form. The archive must contain a `plugin.json`
        manifest at the root with `name` and `version` fields.
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [slug, version, files]
              properties:
                slug:        { type: string, description: "Unique plugin slug" }
                version:     { type: string, description: "Semver, e.g. 1.0.0" }
                displayName: { type: string }
                summary:     { type: string }
                category:    { type: string }
                changelog:   { type: string }
                tags:        { type: string, description: "Comma-separated" }
                files:
                  type: array
                  items: { type: string, format: binary }
      responses:
        "200":
          description: Plugin + new version
          content:
            application/json:
              schema:
                type: object
                properties:
                  plugin:  { $ref: "#/components/schemas/Plugin" }
                  version: { $ref: "#/components/schemas/PluginVersion" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: Version already exists
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /api/v1/plugins/{slug}:
    parameters:
      - { name: slug, in: path, required: true, schema: { type: string } }
    get:
      tags: [Plugins]
      summary: Get a plugin by slug
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Plugin" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Plugins]
      summary: Soft-delete a plugin (owner)
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/MessageOK" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /api/v1/admin/plugins:
    get:
      tags: [Admin]
      summary: List all plugins (any visibility)
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      parameters:
        - { name: visibility, in: query, schema: { type: string, enum: [public, private, pending] } }
        - { name: cursor,     in: query, schema: { type: string } }
        - { name: limit,      in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        "200":
          description: Plugins
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:       { type: array, items: { $ref: "#/components/schemas/Plugin" } }
                  nextCursor: { type: string, nullable: true }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Personal or team token from the device flow / `/api/v1/tokens`. Format is
        opaque; treat it as a secret string. Pass as `Authorization: Bearer <token>`.
    CookieAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: Set by `POST /login` and the OAuth callback. Browser-only.
  responses:
    BadRequest:
      description: Validation error.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthorized:
      description: Missing or invalid token.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: Authenticated but not allowed.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Resource not found.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    MessageOK:
      description: Success.
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string }
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
...
    Plugin:
      type: object
      properties:
        id:               { type: string, format: uuid }
        slug:             { type: string }
        displayName:      { type: string, nullable: true }
        summary:          { type: string, nullable: true }
        visibility:       { type: string, enum: [public, private] }
        category:         { type: string }
        tags:             { type: array, items: { type: string } }
        downloads:        { type: integer }
        starsCount:       { type: integer }
        versionsCount:    { type: integer }
        ownerHandle:      { type: string }
        latestVersionId:  { type: string, format: uuid, nullable: true }
        version:          { type: string, description: "Resolved immutable latest version" }
        fingerprint:      { type: string, description: "Resolved version package SHA-256" }
        sha256Hash:       { type: string, description: "Resolved version package SHA-256" }
        files:
          type: array
          items:
            type: object
            required: [path, size, sha256]
            properties:
              path:   { type: string }
              size:   { type: integer, format: int64 }
              sha256: { type: string }
        moderationStatus: { type: string }
        createdAt:        { type: string, format: date-time }
        updatedAt:        { type: string, format: date-time }
    PluginVersion:
      type: object
      properties:
        id:          { type: string, format: uuid }
        pluginId:    { type: string, format: uuid }
        version:     { type: string, description: "Semver" }
        fingerprint: { type: string, description: "SHA-256 hex" }
        changelog:   { type: string, nullable: true }
        sha256Hash:  { type: string }
        manifest:    { type: object, additionalProperties: true, description: "Parsed plugin.json" }
        files:       { type: array, items: { type: object, properties: { path: { type: string }, size: { type: integer }, sha256: { type: string } } } }
        createdBy:   { type: string, format: uuid }
        yankedAt:    { type: string, format: date-time, nullable: true }
        yankReason:  { type: string, nullable: true }
        createdAt:   { type: string, format: date-time }
...